+++ examples/src/main/scala/zio/config/examples/ErrorAccumulation.scala
@@ -16,7 +16,6 @@ object ErrorAccumulation extends App {
 
   val parsed = runtime.unsafeRun(programWithInvalidSource)
 
-  println(parsed)
   assert(
     parsed ==
       Left(
